migrations/pyproject.toml
[tool.poetry]
name = "molbubi-migrations"
version = "0.1.0"
description = "Database migrations for molbubi.info"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.12"
alembic = "^1.16.4"
sqlalchemy = "^2.0.41"
psycopg2-binary = "^2.9.10"
pydantic-settings = "^2.10.1"
python-dotenv = "^1.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.0.0"
pytest-mock = "^3.12.0"
pytest-asyncio = "^0.23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"